requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
openai>=1.30.0
httpx>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.0
python-dotenv>=1.0.0
